Show messages to user and give an "admire" option
Give the user some helpful messages explaining how to cut the snowflake
and offering some encouragement.

The "admire" button shows the user a large version of the unfolded
snowflake. It shows it as an image rather than a canvas so they can drag
it to their desktop to save.

Also added a "reset" button that removes all cutting.

Worked a bit on styling.

Added short README.md.